+++ src/__test__/unit/vitest.config.ts
@@ -9,7 +9,10 @@ export default defineConfig({
     test: {
         include: ['src/__test__/unit/**/*.test.ts'], // Relative to the project root
         environment: 'happy-dom',
-        reporters: [process.env.CI ? 'junit' : 'verbose'],
-        outputFile: 'test-results/.unit-test-results.xml',
+        // Only set reporters if not provided via CLI (e.g., by IDE)
+        ...(process.env.CI && {
+            reporters: ['junit'],
+            outputFile: 'test-results/.unit-test-results.xml',
+        }),
     },
 });