+++ src/index.ts
@@ -4,3 +4,4 @@ export * from './models/BasicPaymentItem';
 export * from './MaskingUtil';
 export * from './Session';
 export * from './PaymentRequest';
+export * from './CreditCardTokenRequest';